wip: gtk2: insert disk menu option, actions
As well as a new "Insert Disk" menu entry, there are accelerator-only
entries to replace the keyboard shortcuts for inserting into particular
drives.  There has to be a better way of doing the callbacks...

An extra "combo" box is added to the open dialog for selecting which
drive to use.