logging: add log framework, hexdump log type
Might be overkill, but added a concept of log handles.  When a
particular common type of (probably buffered) logging is required,
it can open a log handle of the particular type, and that handle
will keep all the context required.

The first type is LOG_HEXDUMP, which accumulates bytes into a buffer
and emits them in a user-friendly format 16-bytes at a time.