Big joystick interface rewrite
Rather than have keyboard-based virtual joysticks be a special case,
they are now one interface that a joystick module can provide.

Arbitrary joysticks are configured by name like machines and carts,
and only one axis or button is configured at a time.  Standard
configured joysticks are "joy0" and "joy1" for the first and second
actual physical joysticks, and "kjoy0" for whatever the UI module
provides as a "keyboard" interface.

To support the Ctrl+J cycling, a designated "virtual joystick"
configuration is assigned, though at the moment this is always
whatever joystick is named "kjoy0".